+++ src/screen.rs
@@ -4,7 +4,7 @@ use pixen::Image;
 use crate::adb;
 use crate::utils::rgba_into_rgb;
 
-pub(crate) const CHANNELS: usize = 3;
+pub(crate) const CHANNELS: u8 = 3;
 
 pub(super) static SCREENSHOT: RwLock<Option<Image>> =  RwLock::new(None);
 
@@ -19,8 +19,8 @@ pub(super) fn set() {
 
         let image = Image::new(
             rgba_into_rgb(rgba_bytes),
-            w as usize,
-            h as usize,
+            w,
+            h,
             CHANNELS
         );
